started basic strings header, vulkan initialized + debugger, still have to do physical and logical devices.

+++ Renderer/src/game_types.h
@@ -9,10 +9,10 @@ typedef struct game {
     b8      (*initialize)(struct game* game_inst);
 
     // Logic to update the game / simulation
-    b8      (*update)(struct game* game_inst, f32 delta_time);
+    b8      (*update)(struct game* game_inst, f64 delta_time);
 
     // Logic to render the view of the game
-    b8      (*render)(struct game* game_inst, f32 delta_time);
+    b8      (*render)(struct game* game_inst, f64 delta_time);
 
     // Logic to handle window resize is window is a concept of the platform
     void    (*on_resize)(struct game* game_inst, i32 new_width, i32 new_height);
