+++ Testbed/src/game.c
@@ -8,12 +8,12 @@ b8 game_initialize(game* game_inst) {
 }
 
 // Logic to update the game / simulation
-b8 game_update(game* game_inst, f32 delta_time) {
+b8 game_update(game* game_inst, f64 delta_time) {
     return TRUE;
 }
 
 // Logic to render the view of the game
-b8 game_render(game* game_inst, f32 delta_time) {
+b8 game_render(game* game_inst, f64 delta_time) {
     return TRUE;
 }
